requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==41.0.7
cuid2==2.0.0
daphne==4.1.2
dj-database-url==2.1.0
Django==5.0
django-celery-beat==2.6.0
django-charid-field==0.4
django-cloudinary-storage==0.3.0
django-filter==24.2
django-timezone-field==7.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
josepy==1.14.0
kombu==5.3.7
mozilla-django-oidc==3.0.0
msgpack==1.0.8
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pyOpenSSL==23.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.4
requests==2.31.0
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4
